backend/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
pydantic
pinecone
langchain
langchain-openai
langchain-community
beautifulsoup4
requests
pypdf
langchain-google-genai
tavily-python
langgraph




